requirements.txt
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
streamlit-extras>=0.3.0

google-generativeai>=0.3.0
openai>=1.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
torch>=2.0.0
huggingface-hub>=0.17.0

sqlalchemy>=2.0.0

pandas>=2.0.0
numpy>=1.24.0
python-dateutil>=2.8.2

plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
loguru>=0.7.0

pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
passlib[bcrypt]